Split workspace integration parity gates
